correções service e repository

+++ src/main/java/com/sistemaRegistroVerificacao/model/repository/cargo/CargoRepository.java
@@ -4,6 +4,6 @@ import com.sistemaRegistroVerificacao.model.entity.cargo.Cargo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface CargoRepository extends JpaRepository<Cargo, Long>, JpaSpecificationExecutor<Cargo> {
+public interface CargoRepository extends JpaRepository<Cargo, Integer>, JpaSpecificationExecutor<Cargo> {
 }
