+++ src/main/java/com/sistemaRegistroVerificacao/model/repository/ocorrencia/OcorrenciaRepository.java
@@ -4,6 +4,6 @@ import com.sistemaRegistroVerificacao.model.entity.ocorrencia.Ocorrencia;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface OcorrenciaRepository extends JpaRepository<Ocorrencia, Long>, JpaSpecificationExecutor<Ocorrencia> {
+public interface OcorrenciaRepository extends JpaRepository<Ocorrencia, Integer>, JpaSpecificationExecutor<Ocorrencia> {
 }
