+++ src/main/java/com/sistemaRegistroVerificacao/model/repository/statusUsuario/StatusUsuarioRepository.java
@@ -4,6 +4,6 @@ import com.sistemaRegistroVerificacao.model.entity.statusUsuario.StatusUsuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface StatusUsuarioRepository extends JpaRepository<StatusUsuario, Long>, JpaSpecificationExecutor<StatusUsuario> {
+public interface StatusUsuarioRepository extends JpaRepository<StatusUsuario, Integer>, JpaSpecificationExecutor<StatusUsuario> {
 }
