+++ src/main/java/com/sistemaRegistroVerificacao/model/repository/usuario/UsuarioRepository.java
@@ -4,5 +4,5 @@ import com.sistemaRegistroVerificacao.model.entity.usuario.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface UsuarioRepository extends JpaRepository<Usuario, Long>, JpaSpecificationExecutor<Usuario> {
+public interface UsuarioRepository extends JpaRepository<Usuario, Integer>, JpaSpecificationExecutor<Usuario> {
 }
